backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
openai>=1.50.0
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic>=2.0.0

# Dev dependencies
black==24.3.0
ruff==0.3.4
mypy==1.9.0

